**Phase 3/DC update** The launcher and game now run with a hidden console window. Fixed the launcher not updating files after auto-reopening. Fixed the launcher not doing anything when it is out of date. A message box now pops up if you are not running admin. A message box now pops up when the launcher is out of date, taking you to the dl page for the latest launcher. You can now specify the launcher version and game server in the start.bat file for the login server.
Toons going on and offline is now handled by the UD and not the AI.
The game now teleports your toon away and sends you offline correctly when you X out of the game.
Fixed toons not being taken off of other Toons' friends lists when they go offline.
If a district crashes, it's name becomes available on the DistrictNameManager.
A toon can no longer be added to the online toons list if they are already in the list (FriendsManagerUD).
Mods now show up as blue in the friends list and toon panel, devs show up as orange, regular toons show up as black.
Added variables for the different admin tokens in CIGlobals.
The first person camera has been raised higher.
A little bit of dodgeball work (added snowballs to the center).