**COMPILE ALL** **COMPILE CCOGINVASION**
Converted all .mid files to .ogg using SGM-V2.01 soundfont.
Edited tons of files to use .ogg extension instead of .mid.

C++ tournament music now plays perfectly seamless in a different thread no matter the fps.

Created the Python Cog Tournament Music manager.
The tournament music will change based on the events happening (not done yet, it acts weird).
You can specify the number of tournament songs to load in the config_client.

Workaround for the weird bug when calling flattenMedium on the ttc geometry (just made the trees parent to render).

The spray for the squirting flower on fat toons no longer goes up into the air.

Implemented the new logo!
Changed the loading screen UI color to match the color of the logo.

Fixed the overhead laff meter not showing the correct laff of the toon.
Fixed glitchy rendering of the overhead laff meter.

More work on winter dodgeball.
You can now hit other toons with the snowball and they lose hp.

NPCLookInterval is now a LerpQuatInterval so the NPC doesn't spin all the way around.

New network messages for when a tournament round ends and starts.

Updated the Smart Camera physics to simulate the beta toontown smart camera.

Edited the toon victory dance sound to be as in-sync as possible.